Add files via upload
The info graphics illustrated in the project focuses on analysis and tracking of crime in India. The available open crime data sets are being analysed on the basis of different crime categories. Identification of areas with high, moderate and low crime rates and visualizing high alert and safe zones. It enables and assists the senior police officers in better management of police force. It also keeps track of the crime rate through visualization in 29 different states of India. It Identifes high alert and safe zones. Extracting the frequent crime categories.